zakonchil svoi issue's c uchastnikami - dodelal main menu, napisal logiku info uchastnika, redaktirovaniya uchastnika

+++ CompetitionProject/AddParticipant.cs
@@ -37,7 +37,7 @@ namespace CompetitionProject
                 Region = textBox6.Text,
                 SportClub = textBox7.Text,
                 Rank = textBox8.Text,
-                Weight = int.Parse(textBox9.Text),
+                Weight = double.Parse(textBox9.Text),
                 Email = textBox10.Text
             };
             addParticipant.Participants.Add(newParticipant);
